move logic impls into submodule
